Separated graph-editor code into multiple files.
